fix: separate batched question runtime exports
